requirements.txt
Django==5.0.7
django-filter==24.2
djangorestframework==3.15.2
pillow==10.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.0
tzdata==2024.1
rest-framework-simplejwt==0.0.2
setuptools==71.1.0
PyYAML==6.0.1
PyJWT==2.8.0
drf-yasg==1.21.7
djangorestframework-simplejwt==5.3.1
forex-python==1.8
stripe==10.4.0
